perf: se mejora la eficiencia en el calculo
se modifica el algoritmo de cálculo de fibonacci. se modifica el cabezal de la operación CalcularFibonacci para que reciba por parámetro a los dos valores inmediato anteriores en la sucesión. Se corrige la interfaz y todas las clases donde era invocada la operación (tests y controlador)

issue: #12

+++ tests/FibonacciExcercise.UnitTests/FibonacciControllerTest.cs
@@ -25,7 +25,7 @@ namespace FibonacciExcercise.UnitTests
             //int valorEsperado = fib.CalcularFibonacci(7); 
             int valorEsperado = 13;
             FibonacciController fibCont = new FibonacciController(fib2);
-            ActionResult<int> valorActual = fibCont.Get(7);
+            ActionResult<long> valorActual = fibCont.Get(7);
             Assert.AreEqual(valorEsperado, valorActual.Value); 
         }
     }
